backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
opencv-python==4.8.0.76
numpy==1.24.3
pandas==2.0.3
openpyxl==3.1.2
python-dotenv==1.0.0
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.0.0
gunicorn==21.2.0
